mkinitcpio: compress the image with xz
The compression algorithm did not make any difference in my tests regarding boot times, but it decreases the size of a file meant to be copied at least one, and potentially two, times to the 512MiB vfat partition.

Having current initramfs taking up too much space clashes with the "easy kernel swaping" feature to be implemented.